Case Files
These are all the case files that I used to make my macropad. All the .f3d files are the individual bottom and top parts of the case, while the .3mf is the full case with the PCB board inside the actual case itself. This took me lots of time to do but I tried my best.