Schematic & PCB
This was the first part of the process, making the electronics of the actual macropad. As said in the tutorial, I used a XIAO microcontroller, which is an Arduino. I turned the schematic into a PCB so that I was actually able to see and build the electronics into my case.